Restructure models path
