+++ config/jwt.php
@@ -181,7 +181,7 @@ return [
     |
     | This will determine whether a `prv` claim is automatically added to
     | the token. The purpose of this is to ensure that if you have multiple
-    | authentication models e.g. `App\User` & `App\OtherPerson`, then we
+    | authentication models e.g. `App\Models\User` & `App\OtherPerson`, then we
     | should prevent one authentication request from impersonating another,
     | if 2 tokens happen to have the same id across the 2 different models.
     |
